+++ backend/integration_tests/src/communities/import_group_tests.rs
@@ -236,9 +236,9 @@ fn pending_prizes_transferred_to_community() {
 
 fn init_test_data(env: &mut PocketIc, canister_ids: &CanisterIds, controller: Principal) -> TestData {
     let user1 = client::register_diamond_user(env, canister_ids, controller);
-    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
-    let user3 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
-    let user4 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let user2 = client::register_user(env, canister_ids);
+    let user3 = client::register_user(env, canister_ids);
+    let user4 = client::register_user(env, canister_ids);
 
     let group_name = random_string();
     let community_name = random_string();