+++ backend/integration_tests/src/identity_tests.rs
@@ -1,59 +1,18 @@
 use crate::env::ENV;
-use crate::rng::{random_principal, random_string, random_user_principal};
+use crate::rng::{random_principal, random_string};
 use crate::utils::tick_many;
 use crate::{client, TestEnv};
 use rand::random;
 use serde_bytes::ByteBuf;
 use std::ops::Deref;
-use types::Empty;
-
-#[test]
-fn register_via_identity_canister_succeeds() {
-    let mut wrapper = ENV.deref().get();
-    let TestEnv { env, canister_ids, .. } = wrapper.env();
-
-    let (auth_principal, public_key) = random_user_principal();
-    let session_key = ByteBuf::from(random::<[u8; 32]>().to_vec());
-
-    let create_identity_result = client::identity::happy_path::create_identity(
-        env,
-        auth_principal,
-        canister_ids.identity,
-        public_key.clone(),
-        session_key.clone(),
-    );
-
-    client::identity::happy_path::get_delegation(
-        env,
-        auth_principal,
-        canister_ids.identity,
-        session_key,
-        create_identity_result.expiration,
-    );
-
-    let register_response = client::local_user_index::register_user(
-        env,
-        create_identity_result.principal,
-        canister_ids.local_user_index,
-        &local_user_index_canister::register_user::Args {
-            public_key: create_identity_result.user_key,
-            username: random_string(),
-            referral_code: None,
-        },
-    );
-
-    assert!(matches!(
-        register_response,
-        local_user_index_canister::register_user::Response::Success(_)
-    ));
-}
+use types::{DiamondMembershipPlanDuration, Empty};
 
 #[test]
 fn delegation_signed_successfully() {
     let mut wrapper = ENV.deref().get();
     let TestEnv { env, canister_ids, .. } = wrapper.env();
 
-    let user = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let user = client::local_user_index::happy_path::register_legacy_user(env, canister_ids.local_user_index);
 
     client::identity::happy_path::migrate_legacy_principal(env, user.principal, canister_ids.identity);
 
@@ -78,7 +37,7 @@ fn new_users_synced_to_identity_canister() {
     let user_count = 5usize;
 
     let users: Vec<_> = (0..user_count)
-        .map(|_| client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index))
+        .map(|_| client::local_user_index::happy_path::register_legacy_user(env, canister_ids.local_user_index))
         .collect();
 
     env.tick();
@@ -99,7 +58,8 @@ fn migrate_principal_updates_principal_in_all_canisters() {
         ..
     } = wrapper.env();
 
-    let mut user = client::register_diamond_user(env, canister_ids, *controller);
+    let mut user = client::local_user_index::happy_path::register_legacy_user(env, canister_ids.local_user_index);
+    client::upgrade_user(&user, env, canister_ids, *controller, DiamondMembershipPlanDuration::OneMonth);
 
     let group_id = client::user::happy_path::create_group(env, &user, &random_string(), false, true);
     let community_id = client::user::happy_path::create_community(env, &user, &random_string(), false, vec![random_string()]);
@@ -163,7 +123,7 @@ fn principal_migration_job_migrates_all_principals() {
         ..
     } = wrapper.env();
 
-    let platform_operator = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let platform_operator = client::local_user_index::happy_path::register_legacy_user(env, canister_ids.local_user_index);
 
     env.tick();
 
@@ -171,7 +131,7 @@ fn principal_migration_job_migrates_all_principals() {
         client::identity::happy_path::migrate_legacy_principal(env, platform_operator.principal, canister_ids.identity);
 
     let users: Vec<_> = (0..5)
-        .map(|_| client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index))
+        .map(|_| client::local_user_index::happy_path::register_legacy_user(env, canister_ids.local_user_index))
         .collect();
 
     env.tick();