+++ scripts/verify-release.sh
@@ -4,8 +4,13 @@ SCRIPT=$(readlink -f "$0")
 SCRIPT_DIR=$(dirname "$SCRIPT")
 cd $SCRIPT_DIR/..
 
+./scripts/check-docker-is-running.sh || exit 1
+
 RELEASE_VERSION=$1
 EXPECTED_WASM_HASH=$2
+
+[[ -z "$RELEASE_VERSION" ]] && { echo "Release version not provided" ; exit 1; }
+[[ -z "$EXPECTED_WASM_HASH" ]] && { echo "Expected wasm hash not provided" ; exit 1; }
 
 TAG_ID=$(git tag -l --sort=-version:refname "v${RELEASE_VERSION}-*")
 GIT_COMMIT_ID=$(git rev-list $TAG_ID -1)